adding sync command
