+++ e2e_tests/stack_sync_merged_parent_test.go
@@ -47,7 +47,7 @@ func TestStackSyncMergedParent(t *testing.T) {
 	)
 
 	// Everything up to date now, so this should be a no-op.
-	RequireAv(t, "stack", "sync")
+	RequireAv(t, "sync")
 
 	// We simulate a merge here so that our history looks like:
 	//     main:    X
@@ -92,7 +92,7 @@ func TestStackSyncMergedParent(t *testing.T) {
 		"squash commit of stack-1 should not be an ancestor of HEAD of stack-1 before running sync",
 	)
 
-	RequireAv(t, "stack", "sync")
+	RequireAv(t, "sync")
 
 	assert.Equal(t,
 		meta.BranchState{
