+++ e2e_tests/stack_sync_trunk_test.go
@@ -32,7 +32,7 @@ func TestStackSyncTrunk(t *testing.T) {
 	)
 
 	// Everything up to date now, so this should be a no-op.
-	RequireAv(t, "stack", "sync")
+	RequireAv(t, "sync")
 
 	// We simulate a merge here so that our history looks like:
 	//     main:    X --------------> 1S -> 3a
@@ -73,7 +73,7 @@ func TestStackSyncTrunk(t *testing.T) {
 		"squash commit of stack-1 should not be an ancestor of HEAD of stack-2 before running sync",
 	)
 
-	RequireAv(t, "stack", "sync")
+	RequireAv(t, "sync")
 	// At this point, the stack should be:
 	//
 	//     main:    X --------------> 1S -> 3a
